Update creds.py
None of the examples work as the creds variable names don't match the variable names in this creds file.  Updated the variable names here to match the demo scripts.  If you like the names that were there to begin with, can update all the scripts, but this was the smaller change to get it all working.